+++ src/install.js
@@ -1,0 +1,4 @@
+import sagaPlugin from './saga'
+import { activatePlugin } from 'kea'
+
+activatePlugin(sagaPlugin)